Better way to handle Qt AWT backend debugging and update FAQ.awt
	* libraries/clib/awt/qt/evt.cc:
	Safer way to handler processEvent.

	* libraries/clib/awt/qt/img.cc:
	Add type conversion with threshold.

	* libraries/clib/awt/qt/tlk.cc:
	Implement message handler for Qt AWT backend, and it is more
	convenient to debug or catch some messages from Qt core.

	* FAQ/FAQ.awt:
	Update Qt AWT backend debugging information.

+++ libraries/clib/awt/qt/img.cc
@@ -91,7 +91,7 @@ needsFullAlpha ( Toolkit* X, Image *img, double threshold )
 
   if ( !img->alpha ) return 0;
 
-  max = (img->width * img->height) * threshold;
+  max = (img->width * img->height) * (int)threshold;
 
   for ( i=0; i<img->height; i++ ) {
 	for ( j=0; j<img->width; j++ ) {
